requirements.txt
molo.core==3.2.3
molo.surveys==1.0.0
molo.profiles==1.0.1
molo.commenting==0.4.1
molo.polls==2.0.3
gunicorn
psycopg2
django-extensions
molo.usermetadata==1.0.4
six>=1.9
